Crate minimint_wallet

Modules

bitcoincore_rpc

bitcoind

config

db

keys

tweakable

txoproof

Structs

Feerate

PegOut

PegOutSignatureItem

PendingPegOut

PendingTransaction

A peg-out tx that is ready to be broadcast with a tweak for the change UTXO

RoundConsensus

RoundConsensusItem

SpendableUTXO

StatelessWallet 🔒

UnsignedTransaction

A PSBT that is awaiting enough signatures from the federation to becoming a PendingTransaction

UnzipWalletConsensusItem

Wallet

Enums

ProcessPegOutSigError

WalletConsensusItem

WalletError

Constants

CONFIRMATION_TARGET

MIN_PEG_OUT_URGENCY

The urgency of doing a peg-out is defined as the sum over all pending peg-outs of the amount of BTC blocks that have been mined since the peg-out was created. E.g. 10 transactions, each waiting for 10 blocks, would cross a minimum urgency threshold of 100.

Traits

IterUnzipWalletConsensusItem

Functions

broadcast_pending_tx

is_address_valid_for_network

proprietary_tweak_key 🔒

run_broadcast_pending_tx

Type Definitions

PartialSig

PegInDescriptor